glTF: Cleanup includes and defines, split PackedSceneGLTF to own file
