+++ modules/gltf/gltf_animation.h
@@ -71,4 +71,5 @@ private:
 	bool loop = false;
 	Map<int, Track> tracks;
 };
+
 #endif // GLTF_ANIMATION_H
