+++ modules/gltf/gltf_document.h
@@ -31,9 +31,8 @@
 #ifndef GLTF_DOCUMENT_H
 #define GLTF_DOCUMENT_H
 
-#include "gltf_animation.h"
-#include "scene/2d/node_2d.h"
 #include "scene/3d/bone_attachment.h"
+#include "scene/3d/camera.h"
 #include "scene/3d/light.h"
 #include "scene/3d/mesh_instance.h"
 #include "scene/3d/skeleton.h"
@@ -42,6 +41,8 @@
 #include "scene/resources/material.h"
 #include "scene/resources/texture.h"
 
+#include "gltf_animation.h"
+
 #include "modules/modules_enabled.gen.h" // For csg, gridmap.
 
 class GLTFState;
@@ -49,9 +50,14 @@ class GLTFSkin;
 class GLTFNode;
 class GLTFSpecGloss;
 class GLTFSkeleton;
-class CSGShape;
-class GridMap;
 class MultiMeshInstance;
+
+#ifdef MODULE_CSG_ENABLED
+class CSGShape;
+#endif // MODULE_CSG_ENABLED
+#ifdef MODULE_GRIDMAP_ENABLED
+class GridMap;
+#endif // MODULE_GRIDMAP_ENABLED
 
 using GLTFAccessorIndex = int;
 using GLTFAnimationIndex = int;
