+++ modules/gltf/gltf_light.h
@@ -31,7 +31,6 @@
 #ifndef GLTF_LIGHT_H
 #define GLTF_LIGHT_H
 
-#include "core/engine.h"
 #include "core/resource.h"
 
 class GLTFLight : public Resource {
