+++ modules/gltf/gltf_mesh.h
@@ -32,7 +32,6 @@
 #define GLTF_MESH_H
 
 #include "core/resource.h"
-#include "editor/import/resource_importer_scene.h"
 #include "scene/resources/mesh.h"
 
 class GLTFMesh : public Resource {
@@ -54,4 +53,5 @@ public:
 	Array get_instance_materials();
 	void set_instance_materials(Array p_instance_materials);
 };
+
 #endif // GLTF_MESH_H
