+++ modules/gltf/gltf_skin.h
@@ -32,6 +32,7 @@
 #define GLTF_SKIN_H
 
 #include "core/resource.h"
+
 #include "gltf_document.h"
 
 class GLTFSkin : public Resource {
@@ -106,4 +107,5 @@ public:
 	Ref<Skin> get_godot_skin();
 	void set_godot_skin(Ref<Skin> p_godot_skin);
 };
+
 #endif // GLTF_SKIN_H
