+++ modules/gltf/gltf_texture.h
@@ -32,6 +32,7 @@
 #define GLTF_TEXTURE_H
 
 #include "core/resource.h"
+
 #include "gltf_document.h"
 
 class GLTFTexture : public Resource {
