+++ modules/gltf/register_types.cpp
@@ -28,11 +28,10 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
+#ifndef _3D_DISABLED
+
 #include "register_types.h"
 
-#include "editor/editor_node.h"
-#include "editor_scene_exporter_gltf_plugin.h"
-#include "editor_scene_importer_gltf.h"
 #include "gltf_accessor.h"
 #include "gltf_animation.h"
 #include "gltf_buffer_view.h"
@@ -46,8 +45,14 @@
 #include "gltf_spec_gloss.h"
 #include "gltf_state.h"
 #include "gltf_texture.h"
+#include "packed_scene_gltf.h"
 
-#ifndef _3D_DISABLED
+#ifdef TOOLS_ENABLED
+#include "editor/editor_node.h"
+#include "editor_scene_exporter_gltf_plugin.h"
+#include "editor_scene_importer_gltf.h"
+#endif
+
 #ifdef TOOLS_ENABLED
 static void _editor_init() {
 	Ref<EditorSceneImporterGLTF> import_gltf;
@@ -55,10 +60,8 @@ static void _editor_init() {
 	ResourceImporterScene::get_singleton()->add_importer(import_gltf);
 }
 #endif
-#endif
 
 void register_gltf_types() {
-#ifndef _3D_DISABLED
 #ifdef TOOLS_ENABLED
 	ClassDB::APIType prev_api = ClassDB::get_current_api();
 	ClassDB::set_current_api(ClassDB::API_EDITOR);
@@ -68,6 +71,7 @@ void register_gltf_types() {
 	ClassDB::set_current_api(prev_api);
 	EditorNode::add_init_callback(_editor_init);
 #endif
+
 	ClassDB::register_class<GLTFSpecGloss>();
 	ClassDB::register_class<GLTFNode>();
 	ClassDB::register_class<GLTFAnimation>();
@@ -81,8 +85,9 @@ void register_gltf_types() {
 	ClassDB::register_class<GLTFState>();
 	ClassDB::register_class<GLTFDocument>();
 	ClassDB::register_class<PackedSceneGLTF>();
-#endif
 }
 
 void unregister_gltf_types() {
 }
+
+#endif // _3D_DISABLED
