+++ modules/gltf/register_types.h
@@ -28,5 +28,14 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
+#ifndef GLTF_REGISTER_TYPES_H
+#define GLTF_REGISTER_TYPES_H
+
+#ifndef _3D_DISABLED
+
 void register_gltf_types();
 void unregister_gltf_types();
+
+#endif // _3D_DISABLED
+
+#endif // GLTF_REGISTER_TYPES_H
